pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Page Analyzer"
requires-python = ">=3.8.1"
dependencies = [
    "flask>=2.0.0",
    "gunicorn>=20.1.0",
    "python-dotenv>=1.0.0",
    "flake8>=6.0.0",
    "psycopg2-binary>=2.9.0",
    "validators>=0.20.0",
    "requests>=2.32.3",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["page_analyzer"]
